Add dotfiles installer script
Adds an initial version of an installer which copies any files in the
repository to a build directory, and processes any replacement tags in
the form of %prog:arg% where "prog" is the program to run and "arg" is
the last argument to supply to the program.  For example "%echo:hi%"
will be substituted with the result of running `echo hi`.  After
all substitutions are made the installer will link the build files into
destination path and report on any conflicts.